requirements.txt
flask==3.0.0
qrcode[pil]==7.4.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
pillow==10.2.0
pyzbar
python-dotenv==1.0.0